+++ modules/GameBoost.Rendering/SpriteBatch.cs
@@ -56,5 +56,6 @@ namespace GameBoost.Rendering
                 _sprites[i] = sprite;
             }
         }
+
     }
 }